tests/wpt/css-tests/css-writing-modes-3_dev/xhtml1print/unicode-bidi-parsing-001.xht

<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
<html xmlns="http://www.w3.org/1999/xhtml"><head><title>CSS Writing Modes: parsing unicode-bidi: normal, embed, bidi-override</title>
<link href="kojiishi@gmail.com" rel="author" title="Koji Ishii" />
<link href="https://drafts.csswg.org/css2/visuren.html#direction" rel="help" />
<link href="https://drafts.csswg.org/css-writing-modes-3/#unicode-bidi" rel="help" />
<meta content="This test asserts the parser and getComputedStyle works correctly for the unicode-bidi: normal, embed, bidi-override." name="assert" />
<meta content="dom" name="flags" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

</head><body><div data-expected="normal" title="Initial value of unicode-bidi"></div>

<div data-expected="normal" style="unicode-bidi: embed; unicode-bidi: normal"></div>
<div data-expected="embed" style="unicode-bidi: embed"></div>
<div data-expected="bidi-override" style="unicode-bidi: bidi-override"></div>

<div style="unicode-bidi: embed">
  <div data-expected="normal" title="unicode-bidi should not inherit"></div>
</div>

<script>
Array.prototype.forEach.call(document.querySelectorAll("[data-expected]"), function (element) {
  test(function () {
    var actual = getComputedStyle(element).unicodeBidi;
    assert_equals(actual, element.dataset.expected);
  }, element.title || element.getAttribute("style"));
});
</script>
</body></html>
